Add main.py. Slight reshuffle of code into main.py
