+++ programs/fibbonaci.asm
@@ -6,9 +6,10 @@
 @fib_loop
     COPY A ACC
     ADD B
-    JUMP_IF_OVERFLOW @set_initial
+    JUMP_IF_OVERFLOW_FLAG @set_initial
     COPY B A
     COPY ACC B
+    JUMP @fib_loop
 
 // Test using DAT
 @set_initial
